crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
qdrant-client>=1.10.1,<2.0.0
sentence-transformers>=2.6.0
tqdm==4.66.1
python-dotenv==1.0.0
lxml==4.9.3 